requirements.txt
python-dotenv
yfinance
groq
phidata
duckduckgo-search
fastapi
uvicorn
packaging
streamlit